scripts/requirements.txt
python-docx==0.8.11
openai==1.14.0
python-dotenv==1.0.1